[compiler] Move general passes from Codegen/Common/ to compiler/Transforms/
These passes often serve as wrappers around upstream MLIR passes or
implement IREE-specific but backend-agnostic transformations. Making
them in Codegen/Common enforce the deps for phases that use the passes.
The revision moves them to `compiler/Transforms/` and add a `README.md`
to avoid the deps, as they can be used by any phases.
