+++ compiler/src/iree/compiler/Codegen/Common/BUILD.bazel
@@ -116,10 +116,8 @@ iree_compiler_cc_library(
         "GenericVectorization.cpp",
         "HoistStaticallyBoundAllocations.cpp",
         "HoistUnrolledVectorExtractInsertSlice.cpp",
-        "IREECodegenAffineExpandIndexOps.cpp",
         "IREECodegenCanonicalizer.cpp",
         "IREECodegenFoldMemRefAliasOps.cpp",
-        "IREECodegenLowerAffine.cpp",
         "IREEComprehensiveBufferize.cpp",
         "IREEExpandStridedMetadata.cpp",
         "IREEInjectAssumeAlignment.cpp",
