+++ compiler/src/iree/compiler/Codegen/Common/CMakeLists.txt
@@ -109,10 +109,8 @@ iree_cc_library(
     "GenericVectorization.cpp"
     "HoistStaticallyBoundAllocations.cpp"
     "HoistUnrolledVectorExtractInsertSlice.cpp"
-    "IREECodegenAffineExpandIndexOps.cpp"
     "IREECodegenCanonicalizer.cpp"
     "IREECodegenFoldMemRefAliasOps.cpp"
-    "IREECodegenLowerAffine.cpp"
     "IREEComprehensiveBufferize.cpp"
     "IREEExpandStridedMetadata.cpp"
     "IREEInjectAssumeAlignment.cpp"
