+++ compiler/src/iree/compiler/Codegen/LLVMCPU/CMakeLists.txt
@@ -169,6 +169,7 @@ iree_cc_library(
     iree::compiler::Dialect::TensorExt::IR
     iree::compiler::Dialect::Util::IR
     iree::compiler::Dialect::Util::Transforms
+    iree::compiler::Transforms
     iree::compiler::Utils
     iree::schemas::cpu_data
     iree::schemas::instruments
