+++ compiler/src/iree/compiler/Codegen/LLVMGPU/CMakeLists.txt
@@ -202,6 +202,7 @@ iree_cc_library(
     iree::compiler::Dialect::Util::Analysis
     iree::compiler::Dialect::Util::IR
     iree::compiler::Dialect::Util::Transforms
+    iree::compiler::Transforms
     iree::compiler::Utils
   PUBLIC
 )
