+++ compiler/src/iree/compiler/Dialect/HAL/Transforms/CMakeLists.txt
@@ -90,6 +90,7 @@ iree_cc_library(
     iree::compiler::Dialect::Util::IR
     iree::compiler::Dialect::Util::Transforms
     iree::compiler::Modules::IO::Parameters::IR::IOParametersDialect
+    iree::compiler::Transforms
     iree::compiler::Utils
     iree::schemas::instruments
     iree::schemas::instruments::dispatch_def_c_fbs
