+++ compiler/src/iree/compiler/Dialect/VM/Transforms/CMakeLists.txt
@@ -68,6 +68,7 @@ iree_cc_library(
     iree::compiler::Dialect::VM::Conversion::StandardToVM
     iree::compiler::Dialect::VM::Conversion::UtilToVM
     iree::compiler::Dialect::VM::IR
+    iree::compiler::Transforms
     iree::compiler::Utils
   PUBLIC
 )
