+++ compiler/src/iree/compiler/Dialect/VMVX/Transforms/CMakeLists.txt
@@ -91,6 +91,7 @@ iree_cc_library(
     iree::compiler::Dialect::VMVX::Conversion::StandardToVMVX
     iree::compiler::Dialect::VMVX::IR
     iree::compiler::Dialect::VMVX::IR::VMVXDialect
+    iree::compiler::Transforms
     iree::compiler::Utils
   PUBLIC
 )
