+++ compiler/src/iree/compiler/Dialect/VMVX/Transforms/Passes.cpp
@@ -14,6 +14,7 @@
 #include "iree/compiler/Dialect/HAL/Transforms/Passes.h"
 #include "iree/compiler/Dialect/LinalgExt/Transforms/Passes.h"
 #include "iree/compiler/Dialect/Util/Transforms/Passes.h"
+#include "iree/compiler/Transforms/Passes.h"
 #include "iree/compiler/Utils/PassUtils.h"
 #include "mlir/Conversion/AffineToStandard/AffineToStandard.h"
 #include "mlir/Conversion/SCFToControlFlow/SCFToControlFlow.h"
@@ -137,7 +138,7 @@ buildVectorVMVXTransformPassPipeline(OpPassManager &variantPassManager) {
 
 static void buildLoopOptimizationVMVXTransformPassPipeline(
     FunctionLikeNest &funcPassManager) {
-  funcPassManager.addPass(createIREECodegenLowerAffinePass)
+  funcPassManager.addPass(createIREELowerAffinePass)
       .addPass(createForOpCanonicalizationPass)
       .addPass(createIREELoopInvariantCodeMotionPass);
 }
