+++ compiler/src/iree/compiler/Tools/BUILD.bazel
@@ -74,6 +74,7 @@ iree_compiler_cc_library(
         "//compiler/src/iree/compiler/Modules/IO/Parameters/Transforms",
         "//compiler/src/iree/compiler/Pipelines",
         "//compiler/src/iree/compiler/Preprocessing:Passes",
+        "//compiler/src/iree/compiler/Transforms",
         "//compiler/src/iree/compiler/Preprocessing/TransformExtensions:PreprocessingExtensions",
         "//llvm-external-projects/iree-dialects:IREELinalgTransformDialect",
         "//llvm-external-projects/iree-dialects:IREELinalgTransformDialectPasses",
