+++ compiler/src/iree/compiler/Transforms/Passes.cpp
@@ -1,0 +1,18 @@
+// Copyright 2026 The IREE Authors
+//
+// Licensed under the Apache License v2.0 with LLVM Exceptions.
+// See https://llvm.org/LICENSE.txt for license information.
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+
+#include "iree/compiler/Transforms/Passes.h"
+#include "mlir/Pass/Pass.h"
+#include "mlir/Pass/PassManager.h"
+
+namespace mlir::iree_compiler {
+
+#define GEN_PASS_REGISTRATION
+#include "iree/compiler/Transforms/Passes.h.inc"
+
+void registerTransformsPasses() { registerPasses(); }
+
+} // namespace mlir::iree_compiler
